% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Composite_fun.R
\name{BayMC.composite}
\alias{BayMC.composite}
\title{Composite Bayesian Multiple Comparisons Function}
\usage{
BayMC.composite(y, npg, pair, p.probs = FALSE)
}
\arguments{
\item{y}{vector of data}

\item{npg}{vector describing the number of observations in each group}

\item{pair}{vector of length two describing pair of interest}

\item{p.probs}{Do you want a table of posterior probabilities for all generated models? Default is FALSE}
}
\description{
This function performs the Baysian Model Selection approach to Multiple Comparisons Problems
}
\examples{
BayMC.composite()
}
