% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_assgn.R
\name{grp.asn}
\alias{grp.asn}
\title{Group Assignment Function}
\usage{
grp.asn(npg, part.vec)
}
\arguments{
\item{npg}{vector of the number of observations in each group}

\item{part.vec}{vector of the partition of interest, as specified by the setparts() function}
}
\description{
This function creates the various group poolings associated with each set partition
}
\examples{
grp.asn()
}
