% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prob_comp.R
\name{prob.compute}
\alias{prob.compute}
\title{Probability Generating Function}
\usage{
prob.compute(y, grpvec)
}
\arguments{
\item{y}{vector of data}

\item{grpvec}{pooling vector of interest as genereated by grp.asn()}
}
\description{
This function generates an approximation to the posterior probability corresponding to a particular partition
}
\examples{
prob.compute()
}
